LoginAndRegisterController populated
